Add proxy location path setting.
This is a workaround for GitHub Pages. GitHub Pages cannot find the proxy files if their location is not specified. If publishing to a different GitHub Pages URL, the string at the top of src/repl/workers/bundler/index.js (proxyLocation) should be changed. If not on GitHub Pages, this variable should be empty.